requirements.txt
nose
pyaml